backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
firebase-admin==6.1.0
biopython==1.81
torch==2.0.0
numpy==1.24.2
web3==6.0.0
python-multipart==0.0.6
